連続ツール実行不具合の改修
